Update README.md
This just the module that you call to capture the screen.  It is not the support code.  As of 6 September 2022 it is working pretty damn good.  It should work for the MSO4104, but I can not be certain as my MSO4104 has not come back from calibration at TMI yet.